Fix Rubocop errors and tests
