+++ lib/openai/http.rb
@@ -1,7 +1,11 @@
 require "event_stream_parser"
 
+require_relative "http_headers"
+
 module OpenAI
   module HTTP
+    include HTTPHeaders
+
     def get(path:)
       parse_jsonl(conn.get(uri(path: path)) do |req|
         req.headers = headers
@@ -76,37 +80,6 @@ module OpenAI
       else
         File.join(@uri_base, @api_version, path)
       end
-    end
-
-    def headers
-      if azure?
-        azure_headers
-      else
-        openai_headers
-      end.merge(extra_headers)
-    end
-
-    def openai_headers
-      {
-        "Content-Type" => "application/json",
-        "Authorization" => "Bearer #{@access_token}",
-        "OpenAI-Organization" => @organization_id
-      }
-    end
-
-    def azure_headers
-      {
-        "Content-Type" => "application/json",
-        "api-key" => @access_token
-      }
-    end
-
-    def extra_headers
-      @extra_headers ||= {}
-    end
-
-    def add_headers(headers)
-      @extra_headers = extra_headers.merge(headers.transform_keys(&:to_s))
     end
 
     def multipart_parameters(parameters)
